+++ src/controller/livroController.js
@@ -1,6 +1,7 @@
 import { autor } from "../models/Autores.js";
 import { editora } from "../models/Editora.js";
 import livro from "../models/Livro.js";
+import Erro404 from "../error/Erro404.js";
 
 class LivroController {
 
@@ -21,7 +22,11 @@ class LivroController {
     try{
       const listaLivro = await livro.findById(id);
 
-      return res.status(200).json(listaLivro);
+      if(listaLivro !== null){
+        return res.status(200).json(listaLivro);
+      }else{
+        next(new Erro404("ID do livro não foi encontrado"));
+      }
     }catch(error){
       next(error);
     }
@@ -33,10 +38,18 @@ class LivroController {
 
     try{
       const listaLivroEncontrado = await livro.find({
-        editor: editora
+        editor: {
+          name: editora
+        }
       });
 
-      return res.status(200).send(listaLivroEncontrado);
+      console.log(listaLivroEncontrado);
+
+      if(listaLivroEncontrado !== null){
+        return res.status(200).json(listaLivroEncontrado);
+      }else{
+        next(new Erro404("Editora não possui nenhum livro cadastrado"));
+      }
     }catch(error){
       next(error);
     }
@@ -75,38 +88,53 @@ class LivroController {
     const idLivro = req.params.id;
 
     try{
+
       const editoraEncontrada = await editora.findById(idEditora);
       const autorEncontrado = await autor.findById(idAutor);
 
-      const livroAtualizado = {
-        ...novoLivro,
-        editor:{
-          ...editoraEncontrada._doc
-        },
-        author: {
-          ...autorEncontrado._doc
+      if(editoraEncontrada !== null & autorEncontrado !== null){
+        const livroAtualizado = {
+          ...novoLivro,
+          editor:{
+            ...editoraEncontrada._doc
+          },
+          author: {
+            ...autorEncontrado._doc
+          }
+        };
+
+        if(await livro.findById(idLivro) !== null){
+          console.log(livro.findByIdAndUpdate(idLivro, livroAtualizado));
+          await livro.findByIdAndUpdate(idLivro, livroAtualizado);
+
+          return res.status(200).json({
+            message: "Livro atualizado"
+          });
+        }else{
+          next(new Erro404("ID do livro não foi encontrado"));
         }
-      };
 
-      await livro.findByIdAndUpdate(idLivro, livroAtualizado);
-
-      return res.status(200).json({
-        message: "Livro atualizado"
-      });
+      }else{
+        next(new Erro404("ID do autor ou editora não foi encontrado"));
+      }
     }catch(error){
       next(error);
     }
-
   }
 
   static async deletaLivro(req,res,next){
     try{
       const id = req.params.id;
-      await livro.findByIdAndDelete(id);
 
-      return res.status(204).json({
-        message: "Livro deletado com sucesso"
-      });
+      if(await livro.findById(id) !== null){
+        await livro.findByIdAndDelete(id);
+
+        return res.status(204).json({
+          message: "Livro deletado com sucesso"
+        });
+      }else{
+        next(new Erro404("ID do livro não foi encontrado"));
+      }
     }catch(error){
       next(error);
     }
